Python Programing
Question 1:
Define Object Oriented Programming Language?
Answer :
An object oriented programming language is a high level programming language based on the object oriented model. Object oriented programming refers to a type of computer programming in which programmers define the data type of a data structure, and also the types of operations that can be applied to the data structure. Many modern programming languages are object oriented, however some older programming languages, such as Pascal, do offer object oriented versions. Examples of object oriented programming languages include Java, C++ and Smalltalk.
Question 2:
List down the Benefits of OOP?
Answer :
Some benefits of OOP:
. It provides a clear modular structure for programs which makes it good for defining abstract datatypes in which implementation details are hidden
. Objects can also be reused within an across applications. The reuse of software also lowers the cost of development. More effort is put into the object oriented analysis and design, which lowers the overall cost of development.
. It makes software easier to maintain. Since the design is modular, part of the system can be updated in case of issues without a need to make large-scale changes
. Reuse also enables faster development. Object oriented programming languages come with rich libraries of objects, and code developed during projects is also reusable in future projects.
Qusetion 3:
Difference between function and method ?
Answer :
A function is a piece of code that is called by name. It can be passed data to operate on i.e. the parameters and can optionally return data (the return value). All data that is passed to a function is explicitly passed.
A method is a piece of code that is called by a name that is associated with an object. In most respects it is identical to a function except for two key differences:
1. A method is implicitly passed the object on which it was called.
2. A method is able to operate on data that is contained within the class.
Question 4:
Define the following terms :
Answer :
1. Class.
In object oriented programming a class is a blueprint for creating objects (a particular data structure), providing initial values for state and implementations of behavior. A class is an entity that determines how an object will behave and what the object will contain. In other words, it is a blueprint or a set of instruction to build a specific type of object.
2. Object.
An object is nothing but a self contained component which consists of methods and properties to make a particular type of data useful. Object determines the behavior of the class. When you send a message to an object you are asking the object to invoke or execute one of its methods.From a programming point of view an object can be a data structure a variable or a function.
3. Attribute.
In Object oriented programming classes and objects have attributes. Attributes are data stored inside a class or instance and represent the state or quality of the class or instance. In short attributes store information about the instance. Also attributes should not be confused with class functions also known as methods.
4.Behavior.
A class behavior determines how an instance of that class operates for example, how it will react if asked to do something by another class or object or if its internal state changes. Behavior is the only way objects can do anything to themselves or have anything done to them.